Rename CAS1 BookingServices to SpaceBookingService
This is for consistency
