+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/cas1/Cas1SpaceBookingCreateService.kt
@@ -26,7 +26,7 @@ class Cas1SpaceBookingCreateService(
   private val cas1PremisesService: Cas1PremisesService,
   private val cas1SpaceBookingRepository: Cas1SpaceBookingRepository,
   private val cas1ApplicationStatusService: Cas1ApplicationStatusService,
-  private val cas1BookingDomainEventService: Cas1BookingDomainEventService,
+  private val cas1BookingDomainEventService: Cas1SpaceBookingDomainEventService,
   private val cas1BookingEmailService: Cas1BookingEmailService,
   private val cas1OutOfServiceBedRepository: Cas1OutOfServiceBedRepository,
   private val cas1OutOfServiceBedService: Cas1OutOfServiceBedService,
