+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/cas1/Cas1SpaceBookingUpdateService.kt
@@ -23,7 +23,7 @@ import java.util.UUID
 class Cas1SpaceBookingUpdateService(
   private val cas1PremisesService: Cas1PremisesService,
   private val cas1SpaceBookingRepository: Cas1SpaceBookingRepository,
-  private val cas1BookingDomainEventService: Cas1BookingDomainEventService,
+  private val cas1BookingDomainEventService: Cas1SpaceBookingDomainEventService,
   private val cas1BookingEmailService: Cas1BookingEmailService,
   private val clock: Clock,
 ) {
